DebugMode: Split up class DebuggingSolverContext as suggested by @kfriedberger in https://github.com/sosy-lab/java-smt/pull/348#discussion_r1585982049.
