+++ src/org/sosy_lab/java_smt/delegate/debugging/DebuggingProverEnvironment.java
@@ -14,7 +14,7 @@ import org.sosy_lab.java_smt.api.ProverEnvironment;
 public class DebuggingProverEnvironment extends DebuggingBasicProverEnvironment<Void>
     implements ProverEnvironment {
   public DebuggingProverEnvironment(
-      BasicProverEnvironment<Void> pDelegate, DebuggingSolverContext pDebugging) {
+      BasicProverEnvironment<Void> pDelegate, DebuggingContext pDebugging) {
     super(pDelegate, pDebugging);
   }
 }
